ref: Move statsPeriodToDays() into utils/duration/
